[Refactor] JoinModal NavBar에서 관리
